Sử dụng state design pattern

+++ Mr.Gimmick/Mr.Gimmick/TwoDimensionObj.cpp
@@ -10,7 +10,7 @@ void TwoDimensionObj::SetSecondValueOfPoint(int secondValueOfPoint)
 	this->point.SetSecondValue(secondValueOfPoint);
 }
 
-void TwoDimensionObj::SetFirstValueOfDimension(int firstValueOfDimension)
+void TwoDimensionObj::SetFirstValueOfDimension(float firstValueOfDimension)
 {
 	this->dimension.SetFirstValue(firstValueOfDimension);
 }
